requirements.txt
Flask-SQLAlchemy
Flask-MySQLdb
flask-restful
flask-bcrypt
